+++ mdb/lib/mysrc/data/api/api.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 import 'package:dio/dio.dart';
-import 'package:mdb/src/data/model/remote/responce/discover_response.dart';
-import 'package:mdb/src/data/model/remote/responce/genres_response.dart';
-import 'package:mdb/src/data/model/remote/responce/popular_movies_responce.dart';
-import 'package:mdb/src/utils/constants.dart';
+import 'package:mdb/mysrc/data/model/remote/responce/discover_response.dart';
+import 'package:mdb/mysrc/data/model/remote/responce/genres_response.dart';
+import 'package:mdb/mysrc/data/model/remote/responce/popular_movies_responce.dart';
+import 'package:mdb/mysrc/utils/constants.dart';
 
 class Api {
   Dio dio = Dio();
