+++ mdb/lib/mysrc/data/repository/movie_repository.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
-import 'package:mdb/src/data/api/api.dart';
-import 'package:mdb/src/data/model/remote/responce/discover_response.dart';
-import 'package:mdb/src/data/model/remote/responce/genres_response.dart';
-import 'package:mdb/src/data/model/remote/responce/popular_movies_responce.dart';
+import 'package:mdb/mysrc/data/api/api.dart';
+import 'package:mdb/mysrc/data/model/remote/responce/discover_response.dart';
+import 'package:mdb/mysrc/data/model/remote/responce/genres_response.dart';
+import 'package:mdb/mysrc/data/model/remote/responce/popular_movies_responce.dart';
 
 class MovieRepository {
   final api = Api();
